Finalize the new menu world format

+++ common/src/main/java/org/vivecraft/client_vr/menuworlds/FakeBlockAccess.java
@@ -45,14 +45,14 @@ public class FakeBlockAccess implements LevelReader {
 	private int zSize;
 	private float ground;
 
-	private int rotation;
+	private float rotation;
 	private boolean rain;
 	private boolean thunder;
 
 	private BiomeManager biomeManager;
 	private DimensionSpecialEffects dimensionInfo;
 	
-	public FakeBlockAccess(int version, long seed, BlockState[] blocks, byte[] skylightmap, byte[] blocklightmap, Biome[] biomemap, short[][] heightmap, int xSize, int ySize, int zSize, int ground, DimensionType dimensionType, boolean isFlat, int rotation, boolean rain, boolean thunder) {
+	public FakeBlockAccess(int version, long seed, BlockState[] blocks, byte[] skylightmap, byte[] blocklightmap, Biome[] biomemap, short[][] heightmap, int xSize, int ySize, int zSize, int ground, DimensionType dimensionType, boolean isFlat, float rotation, boolean rain, boolean thunder) {
 		this.version = version;
 		this.seed = seed;
 		this.blocks = blocks;
@@ -120,7 +120,7 @@ public class FakeBlockAccess implements LevelReader {
 		return seed;
 	}
 
-	public int getRotation() {
+	public float getRotation() {
 		return rotation;
 	}
 	public boolean getRain() {
